LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/gap-017-expected.txt



FAIL .button 1 assert_equals:
<button class="button" data-expected-height="100">
  <div class="item"></div>
  <div class="item"></div>
</button>
height expected 100 but got 80
